[Fix] Delegate 상위뷰에서 지정해주지 않고 중계해주기
