+++ MyKeynote/MyKeynote/Model/Util/ConstantSize.swift
@@ -1,4 +1,3 @@
-import Foundation
 import UIKit
 
 enum ConstantSize {
