+++ MyKeynote/MyKeynote/Model/Util/UIColorExtension.swift
@@ -1,14 +1,5 @@
 import UIKit
 
-extension String {
-    func subString(from: Int, to: Int) -> String {
-        guard from < count, to >= 0, to - from >= 0 else { return "" }
-        let startIndex = self.index(self.startIndex, offsetBy: from)
-        let endIndex = self.index(self.startIndex, offsetBy: to + 1)
-        return String(self[startIndex..<endIndex])
-    }
-}
-
 extension UIColor {
     var redToUInt8: UInt8 {
         return UInt8(self.cgColor.components![0]*255)
