+++ MyKeynote/MyKeynote/View/ContentView/ContentViewFactory.swift
@@ -1,11 +1,11 @@
 import UIKit
 // 뷰는 그리기만 해야되는데 왜 지가 판단?
 class ContentViewFactory {
-    func createContentView(content: Contentable) -> ContentView {
-        if content.type == .square {
-            return SquareContentView(content: content)
-        } else {
-            return ImageContentView(content: content)
-        }
+    func createSquareContentView(content: Contentable) -> SquareContentView {
+        return SquareContentView(content: content)
+    }
+    func createImageContentView(content: Contentable) -> ImageContentView {
+        return ImageContentView(content: content)
+
     }
 }
