+++ MyKeynote/MyKeynote/View/ContentView/ImageContentView.swift
@@ -16,9 +16,9 @@ final class ImageContentView: UIView, ContentView {
     func setFrame() {
     }
     
-    func changeAlphaValue(value: Int) {
+    func changeAlphaValue(alpha: AlphaType) {
         // alpha가 변할 때 border의 투명도도 변하지 않도록 alpha = 이렇게 안하고 backgroundColor 변경
-        backgroundColor = self.backgroundColor?.withAlphaComponent(AlphaType(rawValue: value)!.alphaValue)
+        backgroundColor = self.backgroundColor?.withAlphaComponent(alpha.alphaValue)
     }
     
     func changeBorder(isClicked: Bool) {
