+++ MyKeynote/MyKeynote/View/ContentView/SquareContentView.swift
@@ -4,7 +4,7 @@ import UIKit
 protocol ContentView: UIView {
     var content: any Contentable { get set }
     func setFrame()
-    func changeAlphaValue(value: Int)
+    func changeAlphaValue(alpha: AlphaType)
     func changeBorder(isClicked: Bool)
 }
 
@@ -37,9 +37,9 @@ final class SquareContentView: UIView, ContentView {
         backgroundColor = color
     }
     
-    func changeAlphaValue(value: Int) {
+    func changeAlphaValue(alpha: AlphaType) {
         // alpha가 변할 때 border의 투명도도 변하지 않도록 alpha = 이렇게 안하고 backgroundColor 변경
-        backgroundColor = self.backgroundColor?.withAlphaComponent(AlphaType(rawValue: value)!.alphaValue)
+        backgroundColor = self.backgroundColor?.withAlphaComponent(alpha.alphaValue)
     }
     
     func changeBorder(isClicked: Bool) {
